Обновление
- константы и структура пакета конфиг вынесены в отдельный файл
- исправлено имя файла footer
